taskReset함수 개선

+++ scripts/taskReset.js
@@ -4,12 +4,19 @@
 */
 const taskReset = () => {
   console.log(`새로고침클릭`);
-  const confirmMessage = confirm(`모든 할 일을 삭제하시겠습니까?`); // 안내메세지 출력
-  if (confirmMessage) {
-    localStorage.clear(); // 메세지의 "확인"을 클릭하면 삭제진행
-    console.log(`전체삭제 실행`);
+
+  if (localStorage.getItem(TODO_KEY)) {
+    // 로컬스토리지 내 값이 있다면?
+    const confirmMessage = confirm(`모든 할 일을 삭제하시겠습니까?`);
+    if (confirmMessage) {
+      localStorage.clear(); // 확인버튼 클릭시 전체삭제 진행
+      console.log(`전체삭제 실행`);
+    } else {
+      console.log(`전체삭제 취소`); // 전체삭제 취소
+    }
   } else {
-    console.log(`전체삭제 취소`);
+    // 로컬스토리지 내 값이 없다면?
+    alert("삭제할 할 일이 없습니다!");
   }
   render();
 };
